requirements.txt
python-vlc
requests
tk